requirements.txt
absl-py==2.3.1
astunparse==1.6.3
certifi==2026.1.4
charset-normalizer==3.4.4
efficientnet==1.0.0
flatbuffers==25.12.19
gast==0.7.0
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
idna==3.11
image-classifiers==1.0.0
ImageIO==2.37.2
keras==3.13.0
Keras-Applications==1.0.8
lazy_loader==0.4
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.1.0
networkx==3.6.1
numpy==1.26.4
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pillow==12.1.0
protobuf==4.25.8
Pygments==2.19.2
requests==2.32.5
rich==14.2.0
scikit-image==0.26.0
scipy==1.16.3
segmentation-models==1.0.1
setuptools==80.9.0
six==1.17.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==3.3.0
tifffile==2025.12.20
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
wheel==0.45.1
wrapt==2.0.1
